// client/src/types/protocol.ts
export enum ClientFunctionDef {
  LOGIN_REQ = 0x0001,
  HEARTBEAT_REQ = 0x0002,
  LOGOUT_REQ = 0x0003,
  REGISTER_REQ = 0x0011,
  LIST_PLAYER_REQ = 0x0021,
  CREATE_PLAYER_REQ = 0x0022,
  SELECT_PLAYER_REQ = 0x0023,
  DELETE_PLAYER_REQ = 0x0024,
  CREATE_AGENT_REQ = 0x0031,
  LIST_AGENT_REQ = 0x0032,
  WORD_TEST_REQ = 0x0041,
  WORD_REVIEW_REQ = 0x0042,
  CHAT_MESSAGE_REQ = 0x0051
}

export enum ServerFunctionDef {
  LOGIN_RSP = 0x1001,
  HEARTBEAT_RSP = 0x1002,
  LOGOUT_RSP = 0x1003,
  REGISTER_RSP = 0x1011,
  LIST_PLAYER_RSP = 0x1021,
  CREATE_PLAYER_RSP = 0x1022,
  SELECT_PLAYER_RSP = 0x1023,
  DELETE_PLAYER_RSP = 0x1024,
  CREATE_AGENT_RSP = 0x1031,
  LIST_AGENT_RSP = 0x1032,
  WORD_TEST_RSP = 0x1041,
  WORD_REVIEW_RSP = 0x1042,
  CHAT_MESSAGE_RSP = 0x1051,
  ERROR_RSP = 0x1fff
}